fix: extend main palette
