+++ src/routes/board.tsx
@@ -1,49 +1,27 @@
 import { Stack } from '@mui/material';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
+import { createTheme, ThemeProvider, useTheme } from '@mui/material/styles';
 import { BackButtonGroup } from 'components/board/BackButtonGroup';
 import { TasksGrid } from 'components/board/TasksGrid';
 import { TitleGroup } from 'components/board/TitleGroup';
 
-const theme = createTheme({
-  palette: {
-    primary: {
-      light: '#ffff6e',
-      main: '#cddc39',
-      dark: '#99aa00',
-      contrastText: '#212121',
-    },
-    secondary: {
-      light: '#ff7961',
-      main: '#f44336',
-      dark: '#ba000d',
-      contrastText: '#fff',
-    },
-    neutral: {
-      main: '#64748B',
-      contrastText: '#fff',
-    },
-  },
-});
-
-declare module '@mui/material/styles' {
-  interface Palette {
-    neutral: Palette['primary'];
-  }
-
-  // allow configuration using `createTheme`
-  interface PaletteOptions {
-    neutral?: PaletteOptions['primary'];
-  }
-}
-
-// Update the Button's color prop options
-declare module '@mui/material/Button' {
-  interface ButtonPropsColorOverrides {
-    neutral: true;
-  }
-}
-
 export const Board = () => {
+  const mainTheme = useTheme();
+  const theme = createTheme(mainTheme, {
+    palette: {
+      primary: {
+        light: '#ffff6e',
+        main: '#cddc39',
+        dark: '#99aa00',
+        contrastText: '#212121',
+      },
+      secondary: {
+        light: '#ff7961',
+        main: '#f44336',
+        dark: '#ba000d',
+        contrastText: '#fff',
+      },
+    },
+  });
   return (
     <ThemeProvider theme={theme}>
       <main>
